Moving color theme file
